Fix #59: Remove empty des_old module
This file has been in the repository for years, without any work done on it.
It is actually creating problems, as it is not namespaced,
its module declaration is inferred to be 'des_old',
which could more easily conflict with other symbols.
Being an empty, completely useless module,
it should be acceptable to remove it without a major version bump.