types: Move structs from specific module to types
Those structs were originally defined in specific modules (e.g. rsa)
and globally aliased to a public name in types (formerly ossl_typ).
Since the modules also depended on the public types, there was
a circular dependency: types import the struct to alias it,
and the module containing the struct imports types to use the alias.

The amount of circular dependency ended up confusing the frontend,
and now some new additions might fail (especially if in a `version` / `static if`).

Since those structs have all been deprecated, and the public name is opaque,
they are not needed after OpenSSL 1.1.0 anyway.
And since the `types` module is publicly imported everywhere, moving the struct
definition will simplify the dependency tree, at the expense of making the struct
definition most widely available.